+++ BE/signup12/src/main/java/dev/codesquad/java/signup12/ApiUserController.java
@@ -1,0 +1,18 @@
+package dev.codesquad.java.signup12;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class ApiUserController {
+
+    @Autowired
+    UserRepository userRepository;
+
+    @GetMapping("/")
+    public User test() {
+        Long id = 1L;
+        return userRepository.findById(id).orElseThrow(() -> new IllegalStateException("NO DATA"));
+    }
+}
