+++ BE/signup12/src/main/java/dev/codesquad/java/signup12/UserController.java
@@ -1,18 +1,13 @@
 package dev.codesquad.java.signup12;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
 
-@RestController
+@Controller
 public class UserController {
 
-    @Autowired
-    UserRepository userRepository;
-
-    @GetMapping("/")
-    public User test() {
-        Long id = 1L;
-        return userRepository.findById(id).orElseThrow(() -> new IllegalStateException("NO DATA"));
+    @GetMapping("/test")
+    public String viewUserForm() {
+        return "/form";
     }
 }
